Introduces repomd search
With this we get a Distribution with a bunch of Repository in the database that
you can sync. Which will create, per Distribution/Repository, a gazillion
Package objects with data from primary, appdata and updateinfo.

The user workflow will be as following:

- As an admin I setup Distribution and it's repositories
- As an admin I sync a Distribution's repositories to the database

- As a user I search for Package of a Distributions in it's repositories
- As a user I extend the search to packages build on OBS for this Distribution (not implemented here)

Separating the Distribution "repository" search from searching through OBS will
hopefully make more clear for newbies that enabling extra repos is kind of
dangerous.